+++ internal/rag/extractor/testmain_test.go
@@ -1,0 +1,13 @@
+package extractor
+
+import (
+	"os"
+	"testing"
+
+	chassis "github.com/ai8future/chassis-go"
+)
+
+func TestMain(m *testing.M) {
+	chassis.RequireMajor(4)
+	os.Exit(m.Run())
+}